Rework getContractWasm
