+++ frontend/store/helpers/base64ToArrayBuffer.js
@@ -1,0 +1,7 @@
+export const base64ToArrayBuffer = (base64String) => {
+  const binaryString = window.atob(base64String);
+  const bytes = Uint8Array.from({ length: binaryString.length }, (_, i) =>
+    binaryString.charCodeAt(i),
+  );
+  return bytes.buffer;
+};
